+++ test/test-DEER-Turing-Logistic.jl
@@ -68,7 +68,9 @@ end
 
 function _deer_logistic_turing_density_model()
     return DensityModel(
-        _deer_logistic_regression(_LR_X, _LR_y); hvp=(β, v) -> _hvp_lr(β, v, _LR_X, _LR_y)
+        _deer_logistic_regression(_LR_X, _LR_y);
+        ad_backend=ADTypes.AutoForwardDiff(),
+        hvp=(β, v) -> _hvp_lr(β, v, _LR_X, _LR_y),
     )
 end