+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Removed
+
+- `DynamicPPLExt` no longer requires `ForwardDiff` as a triggering library to load.
+
 ## [0.2.0] - 2026-06-29
 
 ### Changed